Move next meeting date logic to Github Action script
