+++ .github/workflows/get-next-meeting-date.yml
@@ -1,0 +1,38 @@
+name: Get Next Meeting Date
+on:
+  schedule:
+    - cron: "0 * * * *" # Runs every hour
+
+jobs:
+  execute-script:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: lts/*
+          cache: "npm"
+          cache-dependency-path: client/package-lock.json
+
+      - name: Install axios
+        working-directory: ./client
+        run: npm i axios
+
+      - name: Install cheerio
+        working-directory: ./client
+        run: npm i cheerio
+
+      - name: Install fs
+        working-directory: ./client
+        run: npm i fs
+
+      - name: Compile next meeting date script
+        working-directory: ./client/src/scripts
+        run: npx tsc -t es2022 -m nodenext --moduleResolution nodenext next-meeting-date.ts
+
+      - name: Run next meeting date script
+        working-directory: ./client/src/scripts
+        run: node next-meeting-date.js
