+++ client/src/components/pages/database/header.tsx
@@ -1,40 +1,6 @@
 import { FormattedMessage } from "react-intl";
 import "./header.css";
-import * as cheerio from "cheerio";
-import axios from "axios";
-import { useEffect, useState } from "react";
-
-async function scrapeNextMeetingDate(url: string): Promise<Date | null> {
-  const currentDate = new Date();
-  const currentYear = currentDate.getFullYear();
-
-  const response = await axios.get(url);
-  const $ = cheerio.load(response.data);
-
-  const currentYearElement = $("section#content") // Main page content
-    .find(".paragraphs-item-drawers")
-    .first() // Upcoming Hearing Dates
-    .find(
-      `.paragraphs-item-drawer .field.field-label-hidden div:contains('${currentYear}')` // Label element containing the current year
-    )
-    .parentsUntil(".section-drawers") // Lowest common ancestor of the label element and the list of dates
-    .find(".entity .field ul"); // List of dates
-
-  const currentDateStrings = currentYearElement
-    .text()
-    .split("\n")
-    .filter((dateString) => !!dateString && dateString.includes("Voting"))
-    .map((dateString) => dateString.replace(/\(Voting\)/g, "").trim());
-
-  const meetingDates = currentDateStrings.map(
-    (dateString) => new Date(`${dateString}, ${currentYear}`)
-  );
-
-  const nextMeetingDate =
-    meetingDates.find((date) => date > currentDate) ?? null;
-
-  return nextMeetingDate;
-}
+import NEXT_MEETING_DATE from "../../../data/next-meeting-date.json";
 
 // Show the next meeting date if and only if we have one and it is in the future
 const getNextMeetingText = (nextMeeting: Date | null) => {
@@ -57,20 +23,10 @@ const getNextMeetingText = (nextMeeting: Date | null) => {
 };
 
 const Header = () => {
-  const [nextMeetingDate, setNextMeetingDate] = useState(
-    new Date("2000-01-01") // Use a date in the past so nothing is shown unless we get a valid and future date
-  );
-  useEffect(() => {
-    async function fetchNextMeetingDate() {
-      const date = await scrapeNextMeetingDate(
-        "https://www.boston.gov/departments/licensing-board/licensing-board-information-and-members"
-      );
-      if (date) {
-        setNextMeetingDate(date);
-      }
-    }
-    fetchNextMeetingDate();
-  }, []);
+  const nextMeetingDate = NEXT_MEETING_DATE.nextMeetingDate
+    ? new Date(NEXT_MEETING_DATE.nextMeetingDate)
+    : null;
+
   return (
     <header className="database-header">
       <div className="text-container">
