+++ client/src/data/next-meeting-date.json
@@ -1,0 +1,1 @@
+{ "nextMeetingDate": null }
